refactor: simplify runtime orchestration layer (-1,297 lines, -10 files)
Delete PrivilegedOperationsCoordinator (1,246 lines) and replace with
PrivilegedOperationsRouter (630 lines) — pure helper/sudo routing with
no domain logic. Extract guard/decision logic into standalone
ServiceInstallGuard with 24 golden tests proving behavioral equivalence.

Consolidate 8 RuntimeCoordinator extension files into the main file,
reducing from 10 files to 2 (main + RuleCollections).

Audit KanataDaemonManager vs KanataDaemonService confirms the split is
justified: Manager handles registration lifecycle, Service handles
runtime monitoring.
