--- Sources/KeyPathAppKit/Core/PrivilegedOperationsCoordinator.swift
@@ -1,1246 +1,0 @@
-import AppKit
-import Foundation
-import KeyPathCore
-import KeyPathInstallationWizard
-import KeyPathWizardCore
-
-@MainActor
-protocol PrivilegedOperationsCoordinating: WizardPrivilegedOperating {}
-
-/// Coordinates all privileged operations with hybrid approach (helper vs direct sudo)
-///
-/// **Architecture:** This coordinator implements the hybrid strategy from HELPER.md
-/// - DEBUG builds: Use direct sudo (AuthorizationExecuteWithPrivileges)
-/// - RELEASE builds: Prefer privileged helper; fall back to sudo on failure
-///
-/// **Usage:**
-/// ```swift
-/// let coordinator = PrivilegedOperationsCoordinator.shared
-/// try await coordinator.installRequiredRuntimeServices()
-/// ```
-@MainActor
-public final class PrivilegedOperationsCoordinator {
-    private static let serviceGuardLogPrefix = "[ServiceInstallGuard]"
-    private static let serviceInstallThrottle: TimeInterval = 30
-    private static var lastServiceInstallAttempt: Date?
-    private static var lastSMAppApprovalNotice: Date?
-    private static let smAppApprovalNoticeThrottle: TimeInterval = 5
-    // Clean macOS starts can legitimately take >10s because Kanata sleeps for 2s on startup
-    // and may wait up to 10s for the DriverKit virtual keyboard to become ready.
-    private static let kanataReadinessTimeout: TimeInterval = 20
-    private static let kanataReadinessPollIntervalSeconds: TimeInterval = 0.5
-    private static let kanataLaunchctlNotFoundExitCode: Int32 = 113
-    private static let persistentLaunchctlNotFoundThreshold = 3
-    private static let staleRecoveryBypassLimit = 3
-    private static var staleRecoveryAttemptCount = 0
-
-    enum KanataReadinessResult: Equatable {
-        case ready
-        case pendingApproval
-        case staleRegistration
-        case launchctlNotFoundPersistent
-        case tcpPortInUse
-        case timedOut
-
-        var isSuccess: Bool {
-            self == .ready || self == .pendingApproval
-        }
-
-        var failureDescription: String {
-            switch self {
-            case .ready:
-                "Kanata readiness verified"
-            case .pendingApproval:
-                "SMAppService approval pending"
-            case .staleRegistration:
-                "SMAppService registration is enabled but launchd cannot load the service"
-            case .launchctlNotFoundPersistent:
-                "launchctl repeatedly reported the Kanata service as not found"
-            case .tcpPortInUse:
-                "TCP port 37001 is already in use by an existing Kanata runtime"
-            case .timedOut:
-                "Kanata did not become running + TCP responsive within readiness timeout"
-            }
-        }
-    }
-
-    enum InstallGuardDecision: Equatable {
-        case skipNoInstall
-        case skipPendingApproval
-        case throttled(remaining: TimeInterval)
-        case run(reason: String, bypassedThrottle: Bool)
-    }
-
-    #if DEBUG
-        nonisolated(unsafe) static var serviceStateOverride:
-            (() -> KanataDaemonManager.ServiceManagementState)?
-        nonisolated(unsafe) static var installAllServicesOverride: (() async throws -> Void)?
-        nonisolated(unsafe) static var recoverRequiredRuntimeServicesOverride: (() async throws -> Void)?
-        nonisolated(unsafe) static var killExistingKanataProcessesOverride: (() async throws -> Void)?
-        nonisolated(unsafe) static var kanataReadinessOverride:
-            ((String) async -> KanataReadinessResult)?
-
-        static func resetTestingState() {
-            serviceStateOverride = nil
-            installAllServicesOverride = nil
-            recoverRequiredRuntimeServicesOverride = nil
-            killExistingKanataProcessesOverride = nil
-            kanataReadinessOverride = nil
-            lastServiceInstallAttempt = nil
-            lastSMAppApprovalNotice = nil
-            staleRecoveryAttemptCount = 0
-            KanataDaemonManager.registeredButNotLoadedOverride = nil
-            ServiceHealthChecker.runtimeSnapshotOverride = nil
-        }
-    #endif
-
-    // MARK: - Dependencies
-
-    private let helperManager: HelperManager
-    private let subprocessRunner: SubprocessRunner
-    private let kanataDaemonManager: KanataDaemonManager
-    private let serviceHealthChecker: ServiceHealthChecker
-
-    // MARK: - Singleton
-
-    static let shared = PrivilegedOperationsCoordinator()
-
-    init(
-        helperManager: HelperManager = .shared,
-        subprocessRunner: SubprocessRunner = .shared,
-        kanataDaemonManager: KanataDaemonManager = .shared,
-        serviceHealthChecker: ServiceHealthChecker = .shared
-    ) {
-        self.helperManager = helperManager
-        self.subprocessRunner = subprocessRunner
-        self.kanataDaemonManager = kanataDaemonManager
-        self.serviceHealthChecker = serviceHealthChecker
-        AppLogger.shared.log(
-            "🔐 [PrivCoordinator] Initialized with operation mode: \(Self.operationMode)"
-        )
-    }
-
-    // MARK: - Operation Modes
-
-    enum OperationMode {
-        case privilegedHelper // XPC to root daemon (future: Phase 2)
-        case directSudo // AuthorizationExecuteWithPrivileges (current)
-    }
-
-    /// Determine which operation mode to use based on build configuration
-    static var operationMode: OperationMode {
-        #if DEBUG
-            // Debug builds always use direct sudo for easy contributor testing
-            return .directSudo
-        #else
-            // Release builds prefer helper by default; callers will fall back on errors
-            return .privilegedHelper
-        #endif
-    }
-
-    // MARK: - Unified Privileged Operations API
-
-    /// Remove any installed SMJobBless helper and its daemon plist/logs (developer convenience)
-    public func cleanupPrivilegedHelper() async throws {
-        AppLogger.shared.log("🧹 [PrivCoordinator] Cleaning up privileged helper (dev)")
-
-        // Use tolerant chain so cleanup succeeds even if some files are absent
-        let cmd = """
-        /bin/launchctl bootout system/com.keypath.helper || true; \
-        /bin/rm -f /Library/LaunchDaemons/com.keypath.helper.plist || true; \
-        /bin/rm -f /Library/PrivilegedHelperTools/com.keypath.helper || true; \
-        /bin/rm -f /var/log/com.keypath.helper.stdout.log /var/log/com.keypath.helper.stderr.log || true
-        """
-
-        try await sudoExecuteCommand(cmd, description: "Cleanup privileged helper")
-        AppLogger.shared.log("✅ [PrivCoordinator] Cleanup completed")
-    }
-
-    /// Install all LaunchDaemon services with explicit parameters
-    public func installRequiredRuntimeServices() async throws {
-        AppLogger.shared.log("🔐 [PrivCoordinator] Installing required split-runtime services")
-
-        switch Self.operationMode {
-        case .privilegedHelper:
-            do {
-                try await helperManager.installRequiredRuntimeServices()
-            } catch {
-                AppLogger.shared.log("⚠️ [PrivCoordinator] Helper failed (\(error)), falling back to sudo")
-                try await sudoInstallRequiredRuntimeServices()
-            }
-        case .directSudo:
-            try await sudoInstallRequiredRuntimeServices()
-        }
-    }
-
-    private func currentServiceState() async -> KanataDaemonManager.ServiceManagementState {
-        #if DEBUG
-            if let override = Self.serviceStateOverride {
-                return override()
-            }
-        #endif
-        return await kanataDaemonManager.refreshManagementStateInternal()
-    }
-
-    private func runServiceInstall() async throws {
-        #if DEBUG
-            if let override = Self.installAllServicesOverride {
-                try await override()
-                return
-            }
-        #endif
-        try await installRequiredRuntimeServices()
-    }
-
-    private func enforceKanataRuntimePostcondition(after operation: String) async throws {
-        let readiness = await verifyKanataReadinessAfterInstall(context: operation)
-        guard readiness.isSuccess else {
-            throw PrivilegedOperationError.operationFailed(
-                "Kanata postcondition failed after \(operation): \(readiness.failureDescription)"
-            )
-        }
-
-        if readiness == .pendingApproval {
-            AppLogger.shared.warn(
-                "⚠️ [PrivCoordinator] \(operation) completed but Login Items approval is pending"
-            )
-        } else {
-            AppLogger.shared.log(
-                "✅ [PrivCoordinator] \(operation) postcondition verified (running + TCP responding)"
-            )
-        }
-    }
-
-    private func isRegisteredButNotLoaded() async -> Bool {
-        await kanataDaemonManager.isRegisteredButNotLoaded()
-    }
-
-    private static func decideInstallGuard(
-        state: KanataDaemonManager.ServiceManagementState,
-        staleEnabledRegistration: Bool,
-        now: Date,
-        lastAttempt: Date?
-    ) -> InstallGuardDecision {
-        if state == .smappservicePending {
-            return .skipPendingApproval
-        }
-
-        let requiresInstall = state.needsInstallation || state.needsMigration() || staleEnabledRegistration
-        guard requiresInstall else {
-            // Reset stale-recovery counter when no install is needed, so it doesn't
-            // stay "stuck high" across healthy periods and throttle prematurely on
-            // the next stale event.
-            staleRecoveryAttemptCount = 0
-            return .skipNoInstall
-        }
-
-        if staleEnabledRegistration {
-            staleRecoveryAttemptCount += 1
-            if staleRecoveryAttemptCount <= staleRecoveryBypassLimit {
-                return .run(
-                    reason: "stale-enabled-registration (recovery-attempt=\(staleRecoveryAttemptCount))",
-                    bypassedThrottle: true
-                )
-            }
-        } else {
-            staleRecoveryAttemptCount = 0
-        }
-
-        if let lastAttempt,
-           now.timeIntervalSince(lastAttempt) < serviceInstallThrottle
-        {
-            let remaining = serviceInstallThrottle - now.timeIntervalSince(lastAttempt)
-            return .throttled(remaining: remaining)
-        }
-
-        if staleEnabledRegistration {
-            return .run(
-                reason: "stale-enabled-registration (throttle-applied after \(staleRecoveryBypassLimit) bypasses)",
-                bypassedThrottle: false
-            )
-        }
-
-        return .run(reason: "state=\(state.description)", bypassedThrottle: false)
-    }
-
-    /// Recover runtime services after a failed or stale service state.
-    public func recoverRequiredRuntimeServices() async throws {
-        AppLogger.shared.log("🔐 [PrivCoordinator] Recovering runtime services")
-
-        #if DEBUG
-            if let override = Self.recoverRequiredRuntimeServicesOverride {
-                try await override()
-                return
-            }
-        #endif
-
-        // If the Kanata service is completely uninstalled, install everything first.
-        if try await installServicesIfUninstalled(context: "pre-restart") {
-            AppLogger.shared.log(
-                "✅ [PrivCoordinator] Installed services before restart request – skipping restart call"
-            )
-            try await enforceKanataRuntimePostcondition(after: "recoverRequiredRuntimeServices(pre-restart)")
-            return
-        }
-
-        try await killExistingKanataProcessesForServiceRecovery()
-
-        switch Self.operationMode {
-        case .privilegedHelper:
-            do {
-                try await helperManager.recoverRequiredRuntimeServices()
-                AppLogger.shared.log("✅ [PrivCoordinator] Helper successfully restarted services")
-            } catch {
-                AppLogger.shared.log("⚠️ [PrivCoordinator] Helper failed (\(error)), falling back to sudo")
-                try await sudoRestartServices()
-            }
-        case .directSudo:
-            try await sudoRestartServices()
-        }
-
-        // Double-check after restart – helper path cannot install SMAppService jobs.
-        if try await installServicesIfUninstalled(context: "post-restart") {
-            AppLogger.shared.log("✅ [PrivCoordinator] Service installed after restart attempt")
-        }
-
-        try await enforceKanataRuntimePostcondition(after: "recoverRequiredRuntimeServices")
-    }
-
-    private func killExistingKanataProcessesForServiceRecovery() async throws {
-        #if DEBUG
-            if let override = Self.killExistingKanataProcessesOverride {
-                try await override()
-                return
-            }
-        #endif
-
-        AppLogger.shared.log(
-            "🧹 [PrivCoordinator] Clearing existing Kanata processes before service recovery to avoid TCP port collisions"
-        )
-
-        switch Self.operationMode {
-        case .privilegedHelper:
-            do {
-                try await helperManager.killAllKanataProcesses()
-                AppLogger.shared.log("✅ [PrivCoordinator] Helper cleared existing Kanata processes")
-            } catch {
-                AppLogger.shared.log(
-                    "⚠️ [PrivCoordinator] Helper killAllKanataProcesses failed (\(error)); falling back to sudo"
-                )
-                try await sudoKillAllKanata()
-            }
-        case .directSudo:
-            try await sudoKillAllKanata()
-        }
-    }
-
-    /// Installs all LaunchDaemon services via SMAppService when the Kanata daemon is missing.
-    /// - Returns: `true` if installation was performed.
-    @discardableResult
-    public func installServicesIfUninstalled(context: String) async throws -> Bool {
-        let state = await currentServiceState()
-        AppLogger.shared.log("\(Self.serviceGuardLogPrefix) \(context): state=\(state.description)")
-
-        let staleEnabledRegistration: Bool = if state == .smappserviceActive {
-            await isRegisteredButNotLoaded()
-        } else {
-            false
-        }
-
-        if staleEnabledRegistration {
-            AppLogger.shared.log(
-                "\(Self.serviceGuardLogPrefix) \(context): detected stale SMAppService registration (enabled but not loaded)"
-            )
-        }
-
-        let now = Date()
-        let decision = Self.decideInstallGuard(
-            state: state,
-            staleEnabledRegistration: staleEnabledRegistration,
-            now: now,
-            lastAttempt: Self.lastServiceInstallAttempt
-        )
-
-        switch decision {
-        case .skipPendingApproval:
-            Self.notifySMAppServiceApprovalRequired(context: context)
-            AppLogger.shared.log(
-                "\(Self.serviceGuardLogPrefix) \(context): [pending-approval] skipping install/refresh"
-            )
-            return false
-        case .skipNoInstall:
-            AppLogger.shared.log(
-                "\(Self.serviceGuardLogPrefix) \(context): [no-install-needed] skipping install"
-            )
-            return false
-        case let .throttled(remaining):
-            AppLogger.shared.log(
-                "\(Self.serviceGuardLogPrefix) \(context): [normal-throttle] skipping auto-install (\(String(format: "%.1f", remaining))s remaining)"
-            )
-            return false
-        case let .run(reason, bypassedThrottle):
-            if bypassedThrottle {
-                AppLogger.shared.log(
-                    "\(Self.serviceGuardLogPrefix) \(context): [stale-recovery-bypass] running install despite throttle (\(reason))"
-                )
-            } else {
-                AppLogger.shared.log(
-                    "\(Self.serviceGuardLogPrefix) \(context): [install] proceeding (\(reason))"
-                )
-            }
-        }
-
-        if state.needsMigration() {
-            await removeLegacyKanataPlist(reason: context)
-        }
-
-        Self.lastServiceInstallAttempt = now
-        AppLogger.shared.log("\(Self.serviceGuardLogPrefix) \(context): running SMAppService install")
-        try await runServiceInstall()
-        let postInstallState = await currentServiceState()
-        AppLogger.shared.log(
-            "\(Self.serviceGuardLogPrefix) \(context): install complete, new state=\(postInstallState.description)"
-        )
-        return true
-    }
-
-    /// Regenerate service configuration with current settings
-    public func regenerateServiceConfiguration() async throws {
-        AppLogger.shared.log("🔐 [PrivCoordinator] Regenerating service configuration via SMAppService")
-        // Always use SMAppService path for Kanata
-        try await sudoRegenerateConfig()
-        try await enforceKanataRuntimePostcondition(after: "regenerateServiceConfiguration")
-    }
-
-    /// Install newsyslog config for log rotation
-    public func installNewsyslogConfig() async throws {
-        AppLogger.shared.log("🔐 [PrivCoordinator] Installing newsyslog config")
-
-        switch Self.operationMode {
-        case .privilegedHelper:
-            try await helperInstallNewsyslogConfig()
-        case .directSudo:
-            try await sudoInstallNewsyslogConfig()
-        }
-    }
-
-    /// Repair VirtualHID daemon LaunchDaemon services
-    public func repairVHIDDaemonServices() async throws {
-        AppLogger.shared.log("🔐 [PrivCoordinator] Repairing VHID daemon services")
-
-        switch Self.operationMode {
-        case .privilegedHelper:
-            do { try await helperRepairVHIDServices() } catch {
-                AppLogger.shared.log(
-                    "🚨 [PrivCoordinator] FALLBACK: helper repairVHIDDaemonServices failed: \(error.localizedDescription). Using AppleScript/sudo path."
-                )
-                try await sudoRepairVHIDServices()
-            }
-        case .directSudo:
-            try await sudoRepairVHIDServices()
-        }
-    }
-
-    // MARK: VirtualHID Operations
-
-    /// Activate VirtualHID Manager
-    public func activateVirtualHIDManager() async throws {
-        AppLogger.shared.log("🔐 [PrivCoordinator] Activating VirtualHID Manager")
-
-        switch Self.operationMode {
-        case .privilegedHelper:
-            try await helperActivateVHID()
-        case .directSudo:
-            try await sudoActivateVHID()
-        }
-    }
-
-    /// Uninstall all VirtualHID driver versions
-    public func uninstallVirtualHIDDrivers() async throws {
-        AppLogger.shared.log("🔐 [PrivCoordinator] Uninstalling VirtualHID drivers")
-
-        switch Self.operationMode {
-        case .privilegedHelper:
-            try await helperUninstallDrivers()
-        case .directSudo:
-            try await sudoUninstallDrivers()
-        }
-    }
-
-    /// Download and install specific VirtualHID driver version
-    func installVirtualHIDDriver(version: String, downloadURL: String) async throws {
-        AppLogger.shared.log("🔐 [PrivCoordinator] Installing VirtualHID driver v\(version)")
-
-        switch Self.operationMode {
-        case .privilegedHelper:
-            try await helperInstallDriver(version: version, downloadURL: downloadURL)
-        case .directSudo:
-            try await sudoInstallDriver(version: version, downloadURL: downloadURL)
-        }
-    }
-
-    /// Download and install correct VirtualHID driver version (convenience method)
-    /// Uses VHIDDeviceManager to determine the correct version automatically
-    ///
-    /// **Automatic Fallback:** In release mode, attempts helper first, then falls back
-    /// to sudo if helper fails. This handles phantom registrations and XPC issues gracefully.
-    public func downloadAndInstallCorrectVHIDDriver() async throws {
-        AppLogger.shared.log(
-            "🔐 [PrivCoordinator] Downloading and installing correct VHID driver version"
-        )
-
-        switch Self.operationMode {
-        case .privilegedHelper:
-            do {
-                // Try helper first
-                try await helperInstallCorrectDriver()
-                AppLogger.shared.log("✅ [PrivCoordinator] Helper successfully installed driver")
-            } catch {
-                AppLogger.shared.log("⚠️ [PrivCoordinator] Helper failed (\(error)), falling back to sudo")
-                // Automatic fallback to sudo - catches phantom registrations and XPC failures
-                try await sudoInstallCorrectDriver()
-                AppLogger.shared.log("✅ [PrivCoordinator] Sudo fallback successfully installed driver")
-            }
-        case .directSudo:
-            try await sudoInstallCorrectDriver()
-        }
-    }
-
-    // MARK: Process Management Operations
-
-    /// Terminate a process by PID
-    public func terminateProcess(pid: Int32) async throws {
-        try Self.validateProcessID(pid)
-        AppLogger.shared.log("🔐 [PrivCoordinator] Terminating process PID=\(pid)")
-
-        switch Self.operationMode {
-        case .privilegedHelper:
-            try await helperTerminateProcess(pid: pid)
-        case .directSudo:
-            try await sudoTerminateProcess(pid: pid)
-        }
-    }
-
-    /// Kill all Kanata processes
-    public func killAllKanataProcesses() async throws {
-        AppLogger.shared.log("🔐 [PrivCoordinator] Killing all Kanata processes")
-
-        switch Self.operationMode {
-        case .privilegedHelper:
-            try await helperKillAllKanata()
-        case .directSudo:
-            try await sudoKillAllKanata()
-        }
-    }
-
-    // Removed: legacy non-verified restart. Use restartKarabinerDaemonVerified() instead.
-
-    /// Restart Karabiner VirtualHID daemon with verification (kill all + start + verify)
-    /// Returns true if restart succeeded and daemon is healthy, false otherwise
-    public func restartKarabinerDaemonVerified() async throws -> Bool {
-        AppLogger.shared.log("🔐 [PrivCoordinator] Restarting Karabiner daemon (verified)")
-
-        switch Self.operationMode {
-        case .privilegedHelper:
-            return try await helperRestartKarabinerDaemonVerified()
-        case .directSudo:
-            return try await sudoRestartKarabinerDaemonVerified()
-        }
-    }
-
-    // MARK: Generic Execute
-
-    // Note: executeCommand removed for security. All privileged operations
-    // must be explicitly defined. Internal sudoExecuteCommand remains for
-    // implementation of specific operations.
-
-    // MARK: - Privileged Helper Path (Phase 2 - Future Implementation)
-
-    private func helperRegenerateConfig() async throws {
-        AppLogger.shared.log("🔧 [PrivCoordinator] Bypassing helper - using SMAppService path directly")
-        // Always use SMAppService path for Kanata (helper doesn't support SMAppService)
-        try await sudoRegenerateConfig()
-    }
-
-    private func helperInstallNewsyslogConfig() async throws {
-        do {
-            try await helperManager.installNewsyslogConfig()
-        } catch {
-            AppLogger.shared.log(
-                "🚨 [PrivCoordinator] FALLBACK: helper installNewsyslogConfig failed: \(error.localizedDescription). Using AppleScript/sudo path."
-            )
-            try await sudoInstallNewsyslogConfig()
-        }
-    }
-
-    private func helperRepairVHIDServices() async throws {
-        do {
-            try await helperManager.repairVHIDDaemonServices()
-        } catch {
-            AppLogger.shared.log(
-                "🚨 [PrivCoordinator] FALLBACK: helper repairVHIDDaemonServices failed: \(error.localizedDescription). Using AppleScript/sudo path."
-            )
-            try await sudoRepairVHIDServices()
-        }
-    }
-
-    private func removeLegacyKanataPlist(reason: String) async {
-        let path = KanataDaemonManager.legacyPlistPath
-        guard Foundation.FileManager().fileExists(atPath: path) else { return }
-
-        // Validate path is a safe LaunchDaemons plist before interpolating into shell
-        guard path.hasPrefix("/Library/LaunchDaemons/"),
-              path.hasSuffix(".plist"),
-              !path.contains("'"), !path.contains("\\"), !path.contains(";"),
-              !path.contains("&"), !path.contains("|"), !path.contains("`"),
-              !path.contains("$"), !path.contains("\n")
-        else {
-            AppLogger.shared.log("⚠️ [PrivCoordinator] Refusing to remove suspicious legacy plist path: \(path)")
-            return
-        }
-
-        AppLogger.shared.log("🧹 [PrivCoordinator] Removing legacy Kanata plist (reason: \(reason))")
-        let cmd = """
-        /bin/launchctl bootout system/\(KanataDaemonManager.kanataServiceID) 2>/dev/null || true && \
-        /bin/rm -f '\(path)'
-        """
-        do {
-            try await sudoExecuteCommand(cmd, description: "Remove legacy Kanata plist")
-        } catch {
-            AppLogger.shared.log("⚠️ [PrivCoordinator] Failed to remove legacy Kanata plist: \(error)")
-        }
-    }
-
-    private func verifyKanataReadinessAfterInstall(context: String) async -> KanataReadinessResult {
-        #if DEBUG
-            if let override = Self.kanataReadinessOverride {
-                return await override(context)
-            }
-        #endif
-
-        let managementState = await currentServiceState()
-        if managementState == .smappservicePending {
-            AppLogger.shared.log(
-                "ℹ️ [PrivCoordinator] \(context): readiness pending Login Items approval"
-            )
-            return .pendingApproval
-        }
-
-        let staleEnabledOnEntry = if managementState == .smappserviceActive {
-            await isRegisteredButNotLoaded()
-        } else {
-            false
-        }
-        if staleEnabledOnEntry {
-            AppLogger.shared.error(
-                "❌ [PrivCoordinator] \(context): stale SMAppService registration persisted after install"
-            )
-            return .staleRegistration
-        }
-
-        let deadline = Date().addingTimeInterval(Self.kanataReadinessTimeout)
-        var launchctlNotFoundSamples: [Bool] = []
-        var launchctlNotFoundSuppressedInGraceWindow = false
-
-        while Date() < deadline {
-            if Task.isCancelled {
-                AppLogger.shared.warn("⚠️ [PrivCoordinator] \(context): readiness poll cancelled")
-                return .timedOut
-            }
-
-            let now = Date()
-            let remaining = deadline.timeIntervalSince(now)
-            let timeoutMs = max(50, min(300, Int(remaining * 1000)))
-            let wizardManagementState = WizardServiceManagementState(managementState)
-            let runtimeSnapshot = await serviceHealthChecker.checkKanataServiceRuntimeSnapshot(
-                managementState: wizardManagementState,
-                staleEnabledRegistration: false,
-                timeoutMs: timeoutMs
-            )
-
-            if runtimeSnapshot.isRunning, runtimeSnapshot.isResponding,
-               runtimeSnapshot.inputCaptureReady
-            {
-                return .ready
-            }
-
-            let launchctlNotFound =
-                runtimeSnapshot.launchctlExitCode == Self.kanataLaunchctlNotFoundExitCode
-                    && !runtimeSnapshot.isRunning && !runtimeSnapshot.isResponding
-            launchctlNotFoundSamples.append(launchctlNotFound)
-            if launchctlNotFoundSamples.count > 5 {
-                launchctlNotFoundSamples.removeFirst(launchctlNotFoundSamples.count - 5)
-            }
-            let launchctlNotFoundCount = launchctlNotFoundSamples.filter(\.self).count
-
-            if launchctlNotFoundCount >= Self.persistentLaunchctlNotFoundThreshold {
-                if runtimeSnapshot.recentlyRestarted {
-                    if !launchctlNotFoundSuppressedInGraceWindow {
-                        AppLogger.shared.log(
-                            "ℹ️ [PrivCoordinator] \(context): launchctl not-found threshold reached during restart grace; continuing readiness poll"
-                        )
-                        launchctlNotFoundSuppressedInGraceWindow = true
-                    }
-                } else {
-                    AppLogger.shared.error(
-                        "❌ [PrivCoordinator] \(context): launchctl not-found persisted while Kanata remained down"
-                    )
-                    return .launchctlNotFoundPersistent
-                }
-            } else {
-                launchctlNotFoundSuppressedInGraceWindow = false
-            }
-
-            let sleepSeconds = min(
-                Self.kanataReadinessPollIntervalSeconds,
-                max(0, deadline.timeIntervalSince(Date()))
-            )
-            if sleepSeconds > 0 {
-                do {
-                    try await Task.sleep(for: .seconds(sleepSeconds))
-                } catch is CancellationError {
-                    AppLogger.shared.warn("⚠️ [PrivCoordinator] \(context): readiness poll cancelled during sleep")
-                    return .timedOut
-                } catch {
-                    // No-op: if sleep fails for other reasons, continue toward timeout.
-                }
-            }
-        }
-
-        AppLogger.shared.error(
-            "❌ [PrivCoordinator] \(context): timed out waiting for Kanata readiness"
-        )
-
-        if await detectKanataTCPPortConflict() {
-            AppLogger.shared.error(
-                "❌ [PrivCoordinator] \(context): detected TCP port 37001 conflict while Kanata remained down"
-            )
-            return .tcpPortInUse
-        }
-
-        return .timedOut
-    }
-
-    private func detectKanataTCPPortConflict() async -> Bool {
-        let result: ProcessResult
-        do {
-            result = try await subprocessRunner.run(
-                "/usr/sbin/lsof",
-                args: ["-nP", "-iTCP:37001", "-sTCP:LISTEN"],
-                timeout: 2.0
-            )
-        } catch {
-            AppLogger.shared.log(
-                "⚠️ [PrivCoordinator] Failed to probe TCP port 37001 ownership: \(error.localizedDescription)"
-            )
-            return false
-        }
-
-        guard result.exitCode == 0 else { return false }
-        let lines = result.stdout
-            .split(separator: "\n", omittingEmptySubsequences: true)
-            .map(String.init)
-        guard lines.count >= 2 else { return false }
-        let output = result.stdout.lowercased()
-
-        if output.contains("/library/keypath/bin/kanata") {
-            return true
-        }
-        if output.contains("/applications/keypath.app/contents/library/keypath/kanata") {
-            return true
-        }
-        if output.contains("kanata") {
-            return true
-        }
-
-        return false
-    }
-
-    private func helperActivateVHID() async throws {
-        try await helperManager.activateVirtualHIDManager()
-    }
-
-    private func helperUninstallDrivers() async throws {
-        try await helperManager.uninstallVirtualHIDDrivers()
-    }
-
-    private func helperInstallDriver(version: String, downloadURL: String) async throws {
-        try await helperManager.installVirtualHIDDriver(
-            version: version, downloadURL: downloadURL
-        )
-    }
-
-    private func helperInstallCorrectDriver() async throws {
-        let bundledPkgPath = WizardSystemPaths.bundledVHIDDriverPkgPath
-        guard WizardSystemPaths.bundledVHIDDriverPkgExists else {
-            AppLogger.shared.error("❌ [PrivCoordinator] Bundled VHID driver pkg not found at: \(bundledPkgPath)")
-            throw PrivilegedOperationError.operationFailed(
-                "Bundled VHID driver package not found. This indicates a corrupt KeyPath installation."
-            )
-        }
-        AppLogger.shared.log("📦 [PrivCoordinator] Installing bundled VHID driver: \(bundledPkgPath)")
-        try await helperManager.installBundledVHIDDriver(pkgPath: bundledPkgPath)
-    }
-
-    private func helperTerminateProcess(pid: Int32) async throws {
-        try await helperManager.terminateProcess(pid)
-    }
-
-    /// Terminate a process (helper-first; fallback to sudo with explicit logs)
-    func terminateProcess(_ pid: Int32) async throws {
-        try Self.validateProcessID(pid)
-        do {
-            AppLogger.shared.log("🔐 [PrivCoordinator] Helper-first terminate PID=\(pid)")
-            try await helperTerminateProcess(pid: pid)
-        } catch {
-            AppLogger.shared.log(
-                "🚨 [PrivCoordinator] FALLBACK: helper terminateProcess failed for PID=\(pid): \(error.localizedDescription). Using AppleScript/sudo path."
-            )
-            try await sudoTerminateProcess(pid: pid)
-        }
-    }
-
-    private func helperKillAllKanata() async throws {
-        try await helperManager.killAllKanataProcesses()
-    }
-
-    // Removed: legacy helper restart. Verified path must be used.
-
-    // MARK: - VHID Restart Timing Constants
-
-    // ⚠️ RELIABILITY ROLLBACK: If VHID restart becomes flaky after optimization (Nov 2025),
-    // restore these values: verifyTimeout=3.0, verifyInterval=120ms, settleDelay=300ms
-    // The original conservative timings handle slow/loaded systems better.
-    private static let vhidVerifyTimeoutSeconds: Double = 1.5 // Was: 3.0
-    private static let vhidVerifyInterval: Duration = .milliseconds(100) // was: 120ms
-    private static let vhidSettleDelay: Duration = .milliseconds(150) // was: 300ms
-
-    private func helperRestartKarabinerDaemonVerified() async throws -> Bool {
-        AppLogger.shared.log("🔐 [PrivCoordinator] Helper path: verified restart of Karabiner daemon")
-
-        // Snapshot PRE state (using extracted ServiceHealthChecker)
-        let preLoaded = await serviceHealthChecker.isServiceLoaded(
-            serviceID: "com.keypath.karabiner-vhiddaemon"
-        )
-        let preHealth = await serviceHealthChecker.isServiceHealthy(
-            serviceID: "com.keypath.karabiner-vhiddaemon"
-        )
-        AppLogger.shared.log(
-            "🔎 [PrivCoordinator] PRE: vhiddaemon loaded=\(preLoaded), healthy=\(preHealth)"
-        )
-
-        // 1) Kill any running VirtualHIDDevice daemons via helper (root)
-        do {
-            try await helperManager.restartKarabinerDaemon()
-        } catch {
-            AppLogger.shared.log(
-                "⚠️ [PrivCoordinator] Helper kill phase returned error (continuing): \(error.localizedDescription)"
-            )
-        }
-
-        // 2) Ask helper to restart unhealthy services or install if missing
-        AppLogger.shared.log("🔎 [PrivCoordinator] Calling recoverRequiredRuntimeServices helper...")
-        do {
-            try await helperManager.recoverRequiredRuntimeServices()
-            AppLogger.shared.log("🔎 [PrivCoordinator] recoverRequiredRuntimeServices completed successfully")
-        } catch {
-            AppLogger.shared.log(
-                "⚠️ [PrivCoordinator] Helper recoverRequiredRuntimeServices failed: \(error.localizedDescription)"
-            )
-        }
-
-        // 3) Verification loop - optimized from 3s to 1.5s (Nov 2025)
-        AppLogger.shared.log("🔎 [PrivCoordinator] About to start verification loop...")
-        AppLogger.shared.log("🔎 [PrivCoordinator] Starting verification loop (timeout=\(Self.vhidVerifyTimeoutSeconds)s)")
-        let vhidManager = VHIDDeviceManager()
-        let start = Date()
-        var loopCount = 0
-        while Date().timeIntervalSince(start) < Self.vhidVerifyTimeoutSeconds {
-            loopCount += 1
-            AppLogger.shared.log("🔎 [PrivCoordinator] Verification loop iteration \(loopCount), calling detectRunning()...")
-            let isRunning = await vhidManager.detectRunning()
-            AppLogger.shared.log("🔎 [PrivCoordinator] detectRunning() returned: \(isRunning)")
-            if isRunning {
-                AppLogger.shared.log(
-                    "✅ [PrivCoordinator] Verified: VirtualHIDDevice daemon healthy after helper restart"
-                )
-                return true
-            }
-            try await Task.sleep(for: Self.vhidVerifyInterval)
-        }
-        AppLogger.shared.log("🔎 [PrivCoordinator] Verification loop completed after \(loopCount) iterations, timed out")
-
-        // 4) Single post-verify check (removed repair cascade - user can retry if needed)
-        try await Task.sleep(for: Self.vhidSettleDelay)
-        let postLoaded = await serviceHealthChecker.isServiceLoaded(
-            serviceID: "com.keypath.karabiner-vhiddaemon"
-        )
-        let postHealth = await serviceHealthChecker.isServiceHealthy(
-            serviceID: "com.keypath.karabiner-vhiddaemon"
-        )
-        AppLogger.shared.log(
-            "🔎 [PrivCoordinator] POST: vhiddaemon loaded=\(postLoaded), healthy=\(postHealth)"
-        )
-        if await vhidManager.detectRunning() {
-            AppLogger.shared.log("✅ [PrivCoordinator] Verified after settle: daemon healthy")
-            return true
-        }
-
-        AppLogger.shared.log("❌ [PrivCoordinator] Helper verified restart failed (daemon not healthy)")
-        return false
-    }
-
-    // MARK: - Karabiner Conflict Management
-
-    public func disableKarabinerGrabber() async throws {
-        AppLogger.shared.log("🔐 [PrivCoordinator] Disabling Karabiner grabber via helper")
-        try await helperManager.disableKarabinerGrabber()
-    }
-
-    // MARK: - Direct Sudo Path (Current Implementation)
-
-    private func sudoInstallRequiredRuntimeServices() async throws {
-        let success = await ServiceBootstrapper.shared.installAllServices()
-
-        if !success {
-            throw PrivilegedOperationError.installationFailed("Required runtime service installation failed")
-        }
-    }
-
-    /// Recover runtime services
-    /// Uses extracted ServiceBootstrapper
-    private func sudoRestartServices() async throws {
-        let success = await ServiceBootstrapper.shared.recoverRequiredRuntimeServices()
-
-        if !success {
-            throw PrivilegedOperationError.operationFailed("Service restart failed")
-        }
-    }
-
-    /// Regenerate service configuration (SMAppService for Kanata)
-    /// Uses extracted ServiceBootstrapper
-    private func sudoRegenerateConfig() async throws {
-        AppLogger.shared.log("🔧 [PrivCoordinator] Regenerating service configuration via SMAppService")
-        let success = await ServiceBootstrapper.shared.installAllServices()
-
-        if !success {
-            throw PrivilegedOperationError.operationFailed("Service regeneration failed")
-        }
-    }
-
-    /// Install newsyslog config for log rotation
-    /// Uses extracted ServiceBootstrapper
-    private func sudoInstallNewsyslogConfig() async throws {
-        let success = await ServiceBootstrapper.shared.installNewsyslogConfig()
-
-        if !success {
-            throw PrivilegedOperationError.operationFailed("Newsyslog config installation failed")
-        }
-    }
-
-    /// Repair VHID daemon services
-    /// Uses extracted ServiceBootstrapper
-    private func sudoRepairVHIDServices() async throws {
-        let success = await ServiceBootstrapper.shared.repairVHIDDaemonServices()
-
-        if !success {
-            throw PrivilegedOperationError.operationFailed("VHID daemon repair failed")
-        }
-    }
-
-    /// Activate VirtualHID Manager using VHIDDeviceManager
-    private func sudoActivateVHID() async throws {
-        let vhidManager = VHIDDeviceManager()
-        let success = await vhidManager.activateManager()
-
-        if !success {
-            throw PrivilegedOperationError.operationFailed("VirtualHID activation failed")
-        }
-    }
-
-    /// Uninstall VirtualHID drivers using VHIDDeviceManager
-    private func sudoUninstallDrivers() async throws {
-        let vhidManager = VHIDDeviceManager()
-        let success = await vhidManager.uninstallAllDriverVersions()
-
-        if !success {
-            throw PrivilegedOperationError.operationFailed("Driver uninstallation failed")
-        }
-    }
-
-    /// Install VirtualHID driver using VHIDDeviceManager
-    private func sudoInstallDriver(version _: String, downloadURL _: String) async throws {
-        let vhidManager = VHIDDeviceManager()
-        let success = await vhidManager.downloadAndInstallCorrectVersion()
-
-        if !success {
-            throw PrivilegedOperationError.operationFailed("Driver installation failed")
-        }
-    }
-
-    /// Download and install correct VirtualHID driver version using VHIDDeviceManager
-    private func sudoInstallCorrectDriver() async throws {
-        let vhidManager = VHIDDeviceManager()
-        let success = await vhidManager.downloadAndInstallCorrectVersion()
-
-        if !success {
-            throw PrivilegedOperationError.operationFailed("Driver installation failed")
-        }
-    }
-
-    /// Terminate a process using kill command with admin privileges
-    private func sudoTerminateProcess(pid: Int32) async throws {
-        try Self.validateProcessID(pid)
-        // Try SIGTERM first
-        let termCommand = "/bin/kill -TERM \(pid)"
-
-        do {
-            try await sudoExecuteCommand(termCommand, description: "Terminate process \(pid)")
-        } catch {
-            // If SIGTERM fails, try SIGKILL
-            AppLogger.shared.log("⚠️ [PrivCoordinator] SIGTERM failed, trying SIGKILL")
-            let killCommand = "/bin/kill -9 \(pid)"
-            try await sudoExecuteCommand(killCommand, description: "Force kill process \(pid)")
-        }
-    }
-
-    private static func validateProcessID(_ pid: Int32) throws {
-        guard pid > 0 else {
-            if !TestEnvironment.isTestMode {
-                assertionFailure("[PrivCoordinator] Invalid PID for terminateProcess: \(pid)")
-            }
-            throw PrivilegedOperationError.operationFailed("Invalid process ID: \(pid)")
-        }
-    }
-
-    /// Kill all Kanata processes
-    private func sudoKillAllKanata() async throws {
-        let command = "/usr/bin/pkill -f kanata"
-        try await sudoExecuteCommand(command, description: "Kill all Kanata processes")
-    }
-
-    // Removed: legacy sudo restart. Verified path must be used.
-
-    /// Restart Karabiner daemon with verification (kill all + start managed if possible + sustained verify)
-    /// ⚠️ RELIABILITY ROLLBACK: See timing constants at vhidVerifyTimeoutSeconds if this becomes flaky
-    private func sudoRestartKarabinerDaemonVerified() async throws -> Bool {
-        let daemonPath =
-            "/Library/Application Support/org.pqrs/Karabiner-DriverKit-VirtualHIDDevice/Applications/Karabiner-VirtualHIDDevice-Daemon.app/Contents/MacOS/Karabiner-VirtualHIDDevice-Daemon"
-        let vhidLabel = "com.keypath.karabiner-vhiddaemon"
-        let vhidPlist = "/Library/LaunchDaemons/\(vhidLabel).plist"
-
-        // Determine if a LaunchDaemon is installed; prefer managed restart to prevent duplicates
-        let hasService = Foundation.FileManager().fileExists(atPath: vhidPlist)
-        AppLogger.shared.log("🔐 [PrivCoordinator] VHID LaunchDaemon installed: \(hasService)")
-
-        // Log current PIDs before any action (for diagnostics)
-        let beforePIDs = await getDaemonPIDs()
-        AppLogger.shared.log(
-            "🔎 [PrivCoordinator] VHID PIDs before restart: \(beforePIDs.joined(separator: ", "))"
-        )
-
-        var commands: [String] = []
-
-        if hasService {
-            commands.append("""
-            /bin/launchctl bootout system/\(vhidLabel) 2>/dev/null || true
-            /bin/launchctl disable system/\(vhidLabel) 2>/dev/null || true
-            """)
-        }
-
-        // Kill remaining processes (SIGTERM → SIGKILL) - optimized sleep from 0.3 to 0.15
-        AppLogger.shared.log("🔐 [PrivCoordinator] Killing all VirtualHIDDevice daemon processes")
-        commands.append("""
-        /usr/bin/pkill -f "Karabiner-VirtualHIDDevice-Daemon" 2>/dev/null || true
-        /bin/sleep 0.15
-        /usr/bin/pkill -9 -f "Karabiner-VirtualHIDDevice-Daemon" 2>/dev/null || true
-        /bin/sleep 0.15
-        """)
-
-        // Start via kickstart if service exists; otherwise start directly
-        if hasService {
-            AppLogger.shared.log("🔐 [PrivCoordinator] Starting VHID via launchctl kickstart")
-            commands.append("""
-            /bin/launchctl enable system/\(vhidLabel) 2>/dev/null || true
-            /bin/launchctl kickstart -k system/\(vhidLabel)
-            """)
-        } else {
-            AppLogger.shared.log(
-                "🔐 [PrivCoordinator] LaunchDaemon missing - starting VHID by direct exec"
-            )
-            commands.append("""
-            '\(daemonPath)' > /dev/null 2>&1 &
-            """)
-        }
-
-        let batch = PrivilegedCommandRunner.Batch(
-            label: "restart VirtualHID daemon",
-            commands: commands,
-            prompt: "KeyPath needs to restart the VirtualHID daemon."
-        )
-
-        do {
-            try await sudoExecuteBatch(batch)
-        } catch {
-            AppLogger.shared.log("❌ [PrivCoordinator] Restart batch failed: \(error)")
-            return false
-        }
-
-        // Log PIDs after kill, before start (for diagnostics)
-        let afterKillPIDs = await getDaemonPIDs()
-        AppLogger.shared.log(
-            "🔎 [PrivCoordinator] VHID PIDs after kill: \(afterKillPIDs.joined(separator: ", "))"
-        )
-
-        // Verification loop - optimized from 3s to 1.5s (Nov 2025)
-        let vhidManager = VHIDDeviceManager()
-        let startTime = Date()
-        while Date().timeIntervalSince(startTime) < Self.vhidVerifyTimeoutSeconds {
-            if await vhidManager.detectRunning() {
-                AppLogger.shared.log(
-                    "✅ [PrivCoordinator] Restart verified: daemon is healthy (single instance)"
-                )
-                return true
-            }
-            try await Task.sleep(for: Self.vhidVerifyInterval)
-        }
-
-        // Final diagnostics (removed repair cascade - user can retry if needed)
-        let pids = await getDaemonPIDs()
-        AppLogger.shared.log(
-            "🔎 [PrivCoordinator] VHID PIDs after start: \(pids.joined(separator: ", "))"
-        )
-        if pids.isEmpty {
-            AppLogger.shared.log("❌ [PrivCoordinator] Verification failed: daemon not running")
-        } else {
-            AppLogger.shared.log(
-                "❌ [PrivCoordinator] Verification failed: duplicates (\(pids.count)) PIDs=\(pids.joined(separator: ", "))"
-            )
-        }
-        return false
-    }
-
-    /// Execute a shell command with administrator privileges using osascript
-    /// Public method for use by KanataDaemonManager migration
-    public func sudoExecuteCommand(_ command: String, description: String) async throws {
-        AppLogger.shared.log("🔧 [PrivCoordinator] Executing admin command: \(description)")
-        let result = try await AdminCommandExecutorHolder.shared.execute(
-            command: command, description: description
-        )
-
-        if result.exitCode == 0 {
-            AppLogger.shared.log("✅ [PrivCoordinator] Successfully executed: \(description)")
-        } else {
-            AppLogger.shared.log("❌ [PrivCoordinator] Failed to execute: \(description)")
-            AppLogger.shared.log("❌ [PrivCoordinator] Output: \(result.output)")
-            throw PrivilegedOperationError.commandFailed(
-                description: description,
-                exitCode: result.exitCode,
-                output: result.output
-            )
-        }
-    }
-
-    /// Execute a batch of commands with a single admin prompt.
-    private func sudoExecuteBatch(_ batch: PrivilegedCommandRunner.Batch) async throws {
-        AppLogger.shared.log("🔧 [PrivCoordinator] Executing admin batch: \(batch.label)")
-        let result = try await AdminCommandExecutorHolder.shared.execute(batch: batch)
-
-        if result.exitCode == 0 {
-            AppLogger.shared.log("✅ [PrivCoordinator] Successfully executed batch: \(batch.label)")
-        } else {
-            AppLogger.shared.log("❌ [PrivCoordinator] Failed batch: \(batch.label)")
-            AppLogger.shared.log("❌ [PrivCoordinator] Output: \(result.output)")
-            throw PrivilegedOperationError.commandFailed(
-                description: batch.label,
-                exitCode: result.exitCode,
-                output: result.output
-            )
-        }
-    }
-
-    // MARK: - Helper Methods
-
-    /// Escape a string for use in AppleScript
-    private func escapeForAppleScript(_ string: String) -> String {
-        string
-            .replacingOccurrences(of: "\\", with: "\\\\")
-            .replacingOccurrences(of: "\"", with: "\\\"")
-            .replacingOccurrences(of: "\n", with: "\\n")
-            .replacingOccurrences(of: "\r", with: "\\r")
-    }
-
-    /// Helper: current VHID daemon PIDs (best-effort, no throw)
-    private func getDaemonPIDs() async -> [String] {
-        let pids = await subprocessRunner.pgrep("Karabiner-VirtualHIDDevice-Daemon")
-        return pids.map { String($0) }
-    }
-
-    private static func notifySMAppServiceApprovalRequired(context: String) {
-        let now = Date()
-        if let last = lastSMAppApprovalNotice,
-           now.timeIntervalSince(last) < smAppApprovalNoticeThrottle
-        {
-            return
-        }
-        lastSMAppApprovalNotice = now
-        AppLogger.shared.log(
-            "⚠️ \(serviceGuardLogPrefix) \(context): SMAppService pending approval - notifying UI"
-        )
-        NotificationCenter.default.post(name: .smAppServiceApprovalRequired, object: nil)
-    }
-}
-
-#if DEBUG
-    @_spi(ServiceInstallTesting)
-    extension PrivilegedOperationsCoordinator {
-        func _testEnsureServices(context: String) async throws -> Bool {
-            try await installServicesIfUninstalled(context: context)
-        }
-
-        static func _testResetServiceInstallGuard() {
-            serviceStateOverride = nil
-            installAllServicesOverride = nil
-            lastServiceInstallAttempt = nil
-            staleRecoveryAttemptCount = 0
-        }
-
-        static func _testDecideInstallGuard(
-            state: KanataDaemonManager.ServiceManagementState,
-            staleEnabledRegistration: Bool,
-            now: Date = Date(),
-            lastAttempt: Date? = nil
-        ) -> InstallGuardDecision {
-            decideInstallGuard(
-                state: state,
-                staleEnabledRegistration: staleEnabledRegistration,
-                now: now,
-                lastAttempt: lastAttempt
-            )
-        }
-    }
-#endif
-
-extension PrivilegedOperationsCoordinator: PrivilegedOperationsCoordinating {}
-
-// MARK: - Error Types
-
-enum PrivilegedOperationError: LocalizedError {
-    case installationFailed(String)
-    case operationFailed(String)
-    case commandFailed(description: String, exitCode: Int32, output: String)
-    case executionError(description: String, error: Error)
-
-    var errorDescription: String? {
-        switch self {
-        case let .installationFailed(message):
-            "Installation failed: \(message)"
-        case let .operationFailed(message):
-            "Operation failed: \(message)"
-        case let .commandFailed(description, exitCode, output):
-            "Command failed (\(description)): exit code \(exitCode)\nOutput: \(output)"
-        case let .executionError(description, error):
-            "Execution error (\(description)): \(error.localizedDescription)"
-        }
-    }
-}
